completely new raycasting; fisheye fix

+++ GameHeaders/Player.cpp
@@ -47,7 +47,7 @@ void Player::Debug() const {
     printf("_x: %f, _y: %f\n_state: %c\n_angle: %f\n_progress: %f\n", _x, _y, _state, _angle, _progress);
     for (int y = 0; y < _mazeHeight; y++) {
         for (int x = 0; x < _mazeWidth; x++)
-            if (_x == x && _y == y)
+            if ((int) _x == x && (int) _y == y)
                 printf("_");
             else
                 printf("%d", _maze[y][x]);
